% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{apc.data}
\alias{apc.data}
\alias{cov_c}
\alias{cov_p}
\alias{population}
\alias{cases}
\title{Example dataset for APC model}
\usage{
data(apc)
}
\description{
A dataset containing case counts and population numbers in eight age groups for ten years. Each age group consists of five years.
}
\details{
population: matrix of population data

cases: matrix of case counts

cov_p: covariate for period

cov_c: covariate for cohort
}
\keyword{datasets}
