% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cases}
\alias{cases}
\title{Example dataset of cancer mortalities for APC model}
\format{list}
\usage{
data(apc)
}
\description{
A dataset containing cancer mortalities in eight age groups for ten years. Each age group consists of five years.
}
\keyword{datasets}
