% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_apc.R
\name{check.apc}
\alias{check.apc}
\title{Check apc object, whether MCMC has converged}
\usage{
\method{check}{apc}(x, level = 2, auto = FALSE, ...)
}
\arguments{
\item{x}{An apc object}

\item{level}{level of check; 1 uses point point estimation, 2 uses upper C.I.}

\item{auto}{logical; should be TRUE if called automatically from bamp()}
}
\value{
logical; TRUE if check is fine.
}
\description{
Check apc object, whether MCMC has converged
}
