% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh.R
\name{coh}
\alias{coh}
\title{Compute cohort index from age and period index}
\usage{
coh(agegroup, period, noa, periods_per_agegroup)
}
\arguments{
\item{agegroup}{age group indes}

\item{period}{period indes}

\item{noa}{number of age groups in total}

\item{periods_per_agegroup}{periods per agegroup}
}
\value{
cohort index
}
\description{
Compute cohort index from age and period index
}
