% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cov_c}
\alias{cov_c}
\title{Example dataset for APC model: Cohort covariate for cancer mortalities}
\format{list}
\usage{
data(apc)
}
\description{
A dataset containing cancer mortalities in eight age groups for ten years. Each age group consists of five years. 
Covariate data for cohort effect (80 cohorts).
}
\keyword{datasets}
