% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cov_p}
\alias{cov_p}
\title{Example dataset for APC model: Period covariate for cancer mortalities}
\format{list}
\usage{
data(apc)
}
\description{
A dataset containing cancer mortalities in eight age groups for ten years. Each age group consists of five years. 
Covariate data for period effect.
}
\keyword{datasets}
