% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{population}
\alias{population}
\title{Example dataset for APC model: population}
\format{list}
\usage{
data(apc)
}
\description{
A dataset containing population numbers in eight age groups for ten years. Each age group consists of five years.
}
\keyword{datasets}
