% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_apc.R
\name{print.apc}
\alias{print.apc}
\alias{print}
\title{Print apc objects}
\usage{
\method{print}{apc}(x, ...)
}
\arguments{
\item{x}{apc object}

\item{...}{additional arguments will be ignored}
}
\value{
print
}
\description{
Print apc objects
}
