% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamp.R
\name{bamp}
\alias{bamp}
\title{Bayesian Age-Period-Cohort Modeling and Prediction (bamp)}
\usage{
bamp(cases, population, age, period, cohort, overdisp = FALSE,
  period_covariate = NULL, cohort_covariate = NULL, periods_per_agegroup,
  mcmc.options = list(number_of_iterations = 55000, burn_in = 5000, step = 50,
  tuning = 500), hyperpar = list(age = c(1, 5e-04), period = c(1, 5e-04),
  cohort = c(1, 5e-04), overdisp = c(1, 0.05)), dic = TRUE, parallel = TRUE,
  verbose = FALSE)
}
\arguments{
\item{cases}{number of cases}

\item{population}{population number}

\item{age}{prior for age groups ("rw1", "rw2", "rw1+het", "rw2+het", " ")}

\item{period}{prior for periods ("rw1", "rw2", "rw1+het", "rw2+het", " ")}

\item{cohort}{prior for cohorts ("rw1", "rw2", "rw1+het", "rw2+het", " ")}

\item{overdisp}{logical, add overdispersion to model}

\item{period_covariate}{covariate for period}

\item{cohort_covariate}{covariate for cohort}

\item{periods_per_agegroup}{periods per age group}

\item{mcmc.options}{Options for MCMC, see description}

\item{hyperpar}{Hyper parameters, see description}

\item{dic}{logical. If true. DIC will be computed}

\item{parallel}{logical, should computation be done in parallel}

\item{verbose}{verbose mode}
}
\description{
Bayesian age-period-cohort modeling
}
