% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_apc.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Check apc object, whether MCMC has converged}
\usage{
checkConvergence(x, level = 2, auto = FALSE)
}
\arguments{
\item{x}{An apc object}

\item{level}{level of check; 1 uses point point estimation, 2 uses upper C.I.}

\item{auto}{logical; should be TRUE if called automatically from bamp()}
}
\value{
logical; TRUE if check is fine.
}
\description{
bamp uses Gelman and Rubins R to check convergence for all main parameters. 
All parameters should have R<1.1. 
bamp() runs at least four MCM chains (more if parallel is more than four).
level=2 convergence check is not yet implemented.
}
