% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh.R
\name{coh}
\alias{coh}
\title{Compute cohort index from age and period index}
\usage{
coh(agegroup, period, noa, periods_per_agegroup)
}
\arguments{
\item{agegroup}{age group index}

\item{period}{period index}

\item{noa}{number of age groups in total}

\item{periods_per_agegroup}{periods per age group}
}
\value{
cohort index
}
\description{
Compute cohort index from age and period index
}
\examples{
# last agegroup in first period equals first cohort
coh(10, 1, 10, 5)  

# first agegroup in last period equals last cohort 
coh(1, 8, 10, 5) 
}
