% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_apc.R
\name{effects.apc}
\alias{effects.apc}
\title{Effects from Fitted APC Model}
\usage{
\method{effects}{apc}(object, mean = FALSE, quantiles = 0.5, save = FALSE,
  ...)
}
\arguments{
\item{object}{an apc object}

\item{mean}{logical. If TRUE, mean effects are computed}

\item{quantiles}{Scalar or vector of quantiles to compute (only if mean=FALSE)}

\item{save}{logical. If TRUE, the apc object including the effects is returned}

\item{...}{Additional arguments will be ignored}
}
\value{
List of age, period, cohort effects or apc object including effects
}
\description{
Effects from Fitted APC Model
}
