% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_apc.R
\name{plot.apc}
\alias{plot.apc}
\title{Plot apc object}
\usage{
\method{plot}{apc}(x, quantiles = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{x}{apc object}

\item{quantiles}{quantiles to plot}
}
\value{
plot
}
\description{
Plot apc object
}
