% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_apc.R
\name{predict_apc}
\alias{predict_apc}
\title{Prediction for age-period-cohort models}
\usage{
predict_apc(object, periods = 0, population = NULL,
  quantiles = c(0.05, 0.5, 0.95), update = FALSE)
}
\arguments{
\item{object}{apc object}

\item{periods}{number of periods to predict}

\item{population}{matrix of (predicted) population, if NULL, population data from original bamp call will be used}

\item{quantiles}{vector of quantiles to compute}

\item{update}{boolean. If TRUE, object will be returned with results added to the object}
}
\value{
list with quantiles of predicted probabilities (pr), predicted cases (cases) and predicted cases per period (cases_period)
and a list samples with MCMC samples of pr, cases and cases_period. 
If update=TRUE, the apc object will be returned with this list (predicted) added.
}
\description{
Prediction for age-period-cohort models
}
